+++ src/java/com/android/internal/telephony/gsm/GsmServiceStateTracker.java
@@ -1957,7 +1957,7 @@ final class GsmServiceStateTracker extends ServiceStateTracker {
 
         switch (notifyType) {
         case PS_ENABLED:
-            long dataSubId = SubscriptionManager.getDefaultDataSubId();
+            int dataSubId = SubscriptionManager.getDefaultDataSubId();
             if (dataSubId != mPhone.getSubId()) {
                 return;
             }
